fix(core): update header file includes and correct data type in AHBPrescTable

+++ core/embed/trezorhal/stm32wb55/startup_init.c
@@ -24,7 +24,7 @@
 
 #ifdef KERNEL_MODE
 
-const uint8_t AHBPrescTable[16] = {0U, 0U, 0U, 0U, 0U, 0U, 0U, 0U,
+const uint32_t AHBPrescTable[16] = {0U, 0U, 0U, 0U, 0U, 0U, 0U, 0U,
                                    1U, 2U, 3U, 4U, 6U, 7U, 8U, 9U};
 const uint8_t APBPrescTable[8] = {0U, 0U, 0U, 0U, 1U, 2U, 3U, 4U};
 const uint32_t MSIRangeTable[16] = {48000000U, 24000000U, 16000000U, 12000000U,
